move warnings into diagnostics package
