+++ lenskit/lenskit/__init__.py
@@ -17,19 +17,3 @@ try:
 except PackageNotFoundError:
     # package is not installed
     __version__ = "UNKNOWN"
-
-
-class DataWarning(UserWarning):
-    """
-    Warning raised for detectable problems with input data.
-    """
-
-    pass
-
-
-class ConfigWarning(UserWarning):
-    """
-    Warning raised for detectable problems with algorithm configurations.
-    """
-
-    pass
